+++ src/Airship.Net/Platforms/iOS/Airship.cs
@@ -50,6 +50,7 @@ namespace AirshipDotNet
         private readonly IAirshipPrivacyManager _privacyManager;
         private readonly IAirshipFeatureFlagManager _featureFlagManager;
         private readonly IAirshipPreferenceCenter _preferenceCenter;
+        private readonly IAirshipMessageCenter _messageCenter;
 
         public Airship()
         {
@@ -62,6 +63,7 @@ namespace AirshipDotNet
             _privacyManager = new AirshipPrivacyManager(_module);
             _featureFlagManager = new AirshipFeatureFlagManager(_module);
             _preferenceCenter = new AirshipPreferenceCenter(_module);
+            _messageCenter = new AirshipDotNet.Platforms.iOS.Modules.AirshipMessageCenter(_module);
         }
 
         private void Initialize()
@@ -75,7 +77,6 @@ namespace AirshipDotNet
 
             // Note: Push notification status update event is not directly available in SDK 19
             // This functionality may need to be implemented differently using the new SDK APIs
-
         }
 
         /// <summary>
@@ -190,5 +191,6 @@ namespace AirshipDotNet
         public static IAirshipPrivacyManager PrivacyManager => Instance._privacyManager;
         public static IAirshipFeatureFlagManager FeatureFlagManager => Instance._featureFlagManager;
         public static IAirshipPreferenceCenter PreferenceCenter => Instance._preferenceCenter;
+        public static IAirshipMessageCenter MessageCenter => Instance._messageCenter;
     }
 }